+++ src/components/Callout/Callout.component.tsx
@@ -35,7 +35,7 @@ const defaultCalloutProps: CalloutDefaultProps = {
  * A card to display permanent feedback information.
  * Its color indicates the type of feedback.
  *
- * @version 0.0.4
+ * @version 0.0.5
  *
  * @param {CalloutProps & TestIdProps} props - Callout component props
  * @returns {JSX.Element}