+++ src/components/List/context/ListContext.config.ts
@@ -8,23 +8,19 @@ import {
   filterOptionMatch,
   filterPresetData,
   type FilterBooleanOperator,
-  type FilterComparisonOperator,
   type FilterConfig,
   type FilterData,
   type FilterDataMap,
-  type FilterOptionConfig,
   type FilterOptionData,
   type FilterOptionDataOrSignature,
   type FilterOptionDividerData,
   type FilterOptionMap,
   type FilterOptionMatch,
-  type FilterOptionValue,
-  type FilterPresetConfig,
   type FilterPresetData,
   type FilterPresetMap,
-  type FilterProperty,
   type ListConfigAsyncFilterFn,
   type ListConfigAsyncFilterPresetFn,
+  type ListConfigAsyncOptionFn,
   type ListConfigAsyncOptionsFn,
   type ListConfigFilterFn,
   type ListConfigFilterFnParams,
@@ -35,8 +31,6 @@ import {
   type ListConfigMatchFnParams,
   type ListConfigNotFn,
   type ListConfigOptionFn,
-  type ListConfigOptionFnParams,
-  type ListConfigOptionLabeledValue,
   type ListConfigOptionsFn,
   type UseDataProviderFn,
 } from "../modules";
@@ -56,8 +50,20 @@ import type {
   UseListConfigReturn,
 } from "./ListContext.types";
 import type { MaybePromise } from "@types";
-import type { Nullable, VoidFn } from "@ubloimmo/front-util";
+import type { VoidFn } from "@ubloimmo/front-util";
 
+/**
+ * Hook that exposes callbacks that create & register filter options, filters & filter presets asynchronously
+ *
+ * @param {ListConfigMatchFn<TItem>} match - The match function to use
+ * @param {VoidFn<[FilterOptionData<TItem>]>} registerOption - The function to register a filter option
+ * @param {VoidFn<[FilterData]>} registerFilter - The function to register a filter
+ * @param {VoidFn<[FilterData]>} updateFilter - The function to update a filter
+ * @param {VoidFn<[FilterPresetData]>} registerFilterPreset - The function to register a filter preset
+ * @param {VoidFn<[FilterPresetData]>} updateFilterPreset - The function to update a filter preset
+ * @param {HydrateFilterConfigFn} hydrateFilterConfig - The function to hydrate a filter config
+ * @returns {UseListConfigAsyncReturn<TItem>} The created filter options, filters & filter presets
+ */
 const useListConfigAsync: UseListConfigAsync = <TItem extends object>(
   match: ListConfigMatchFn<TItem>,
   registerOption: VoidFn<[FilterOptionData<TItem>]>,
@@ -69,6 +75,39 @@ const useListConfigAsync: UseListConfigAsync = <TItem extends object>(
 ) => {
   const logger = useLogger("ListConfigAsync");
 
+  /**
+   * Creates & registers a filter option from the provided data
+   *
+   * @param {string} label - The label of the generated filter option
+   * @param {ListConfigAsyncOptionMatchGetterFn<TItem>} matchOrMatchesGetter - An async callback that returns the match or matches to create the filter option from
+   * @param {FilterOptionConfig} [config = {}] - The config of the generated filter option, gets merged with the shared config
+   * @returns {Promise<FilterOptionData<TItem>>} The created filter option
+   * @throws If the async callback throws an error
+   */
+  const option = useCallback<ListConfigAsyncOptionFn<TItem>>(
+    async (label, matchOrMatchesGetter, config = {}) => {
+      try {
+        const matchOrMatches = await matchOrMatchesGetter();
+        const optionData = filterOptionData(label, matchOrMatches, config);
+        registerOption(optionData);
+        return optionData;
+      } catch (error) {
+        logger.error(error);
+        throw error;
+      }
+    },
+    [logger, registerOption]
+  );
+
+  /**
+   * Creates & registers filter options from the provided data
+   *
+   * @param {FilterProperty<TItem>} property - The property to create filter options from
+   * @param {FilterComparisonOperator} comparison - The comparison operator to create filter options from
+   * @param {ListConfigAsyncOptionsGetLabeledValuesFn} labeledValuesGetter - An async callback that returns the labeled values to create filter options from
+   * @param {FilterOptionConfig} [sharedConfig = {}] - The shared config of the generated filter options, gets merged with the config of each option
+   * @returns {Promise<FilterOptionData<TItem>[]>} The created filter options or an empty array if an error occurs
+   */
   const options = useCallback<ListConfigAsyncOptionsFn<TItem>>(
     async (property, comparison, labeledValuesGetter, sharedConfig = {}) => {
       try {
@@ -91,6 +130,14 @@ const useListConfigAsync: UseListConfigAsync = <TItem extends object>(
     [logger, match, registerOption]
   );
 
+  /**
+   * Awaits for the provided options or signatures to load
+   *
+   * @template TPayload - The type of the options or signatures
+   *
+   * @param {MaybePromise<MaybePromise<TPayload>[]>} optionsOrSignatures - The options or signatures to await, can be a promise of an array of promises
+   * @returns {Promise<TPayload[]>} The loaded options or signatures
+   */
   const awaitAsyncOptions = useCallback(
     async <TPayload>(
       optionsOrSignatures: MaybePromise<MaybePromise<TPayload>[]>
@@ -100,16 +147,16 @@ const useListConfigAsync: UseListConfigAsync = <TItem extends object>(
     []
   );
 
+  /**
+   * Creates & registers a filter from an array of asynchronous or synchronous options or signatures
+   *
+   * @param {string} label - The label of the generated filter
+   * @param {MaybePromise<MaybePromise<FilterOptionDataOrSignature<TItem> | FilterOptionDividerData>[]>} optionsOrSignaturesPromise - The options or signatures to create the filter from, can be a promise of an array of promises
+   * @param {FilterConfig} [config = {}] - The config of the generated filter, gets merged with the shared config
+   * @returns {Promise<Nullable<FilterData>>} The created filter or null if an error occurs
+   */
   const filter = useCallback<ListConfigAsyncFilterFn<TItem>>(
-    async (
-      label: string,
-      optionsOrSignaturesPromise: MaybePromise<
-        MaybePromise<
-          FilterOptionDataOrSignature<TItem> | FilterOptionDividerData
-        >[]
-      >,
-      config?: FilterConfig
-    ): Promise<Nullable<FilterData>> => {
+    async (label, optionsOrSignaturesPromise, config = {}) => {
       try {
         const hydrated = hydrateFilterConfig(config);
         // store temporary filter data and mark it as loading
@@ -138,14 +185,16 @@ const useListConfigAsync: UseListConfigAsync = <TItem extends object>(
     ]
   );
 
+  /**
+   * Creates & registers a filter preset from an array of asynchronous or synchronous options or signatures
+   *
+   * @param {string} label - The label of the generated filter preset
+   * @param {MaybePromise<MaybePromise<FilterOptionDataOrSignature<TItem>>[]>} optionsOrSignaturesPromise - The options or signatures to create the filter preset from, can be a promise of an array of promises
+   * @param {FilterPresetConfig} [config = {}] - The config of the generated filter preset, gets merged with the shared config
+   * @returns {Promise<Nullable<FilterPresetData>>} The created filter preset or null if an error occurs
+   */
   const filterPreset = useCallback<ListConfigAsyncFilterPresetFn<TItem>>(
-    async (
-      label: string,
-      optionsOrSignaturesPromise: MaybePromise<
-        MaybePromise<FilterOptionDataOrSignature<TItem>>[]
-      >,
-      config?: FilterPresetConfig
-    ): Promise<Nullable<FilterPresetData>> => {
+    async (label, optionsOrSignaturesPromise, config = {}) => {
       try {
         // store temporary filter preset data and mark it as loading
         const tempFilterPresetData = filterPresetData<TItem>(
@@ -176,12 +225,19 @@ const useListConfigAsync: UseListConfigAsync = <TItem extends object>(
   );
 
   return {
+    option,
     options,
     filter,
     filterPreset,
   };
 };
 
+/**
+ * Hook that manages the list config by exposing callbacks that create & register filter options, filters & filter presets
+ *
+ * @param {UseDataProviderFn<TItem>} dataProvider - The data provider to use
+ * @returns {UseListConfigReturn<TItem>} Object containing the list config & callbacks to further configure the list
+ */
 export const useListConfig: UseListConfig = <TItem extends object>(
   dataProvider: UseDataProviderFn<TItem>
 ): UseListConfigReturn<TItem> => {
@@ -298,6 +354,14 @@ export const useListConfig: UseListConfig = <TItem extends object>(
     BooleanOperators.AND
   );
 
+  /**
+   * Creates a match from the provided data
+   *
+   * @param {FilterProperty<TItem>} property - The property to create the match from
+   * @param {FilterComparisonOperator} comparison - The comparison operator to create the match from
+   * @param {FilterOptionValue} value - The value to create the match from
+   * @returns {FilterOptionMatch<TItem>} The created match
+   */
   const match = useCallback<ListConfigMatchFn<TItem>>(
     (...params: ListConfigMatchFnParams<TItem>): FilterOptionMatch<TItem> => {
       return filterOptionMatch(...params);
@@ -305,52 +369,74 @@ export const useListConfig: UseListConfig = <TItem extends object>(
     []
   );
 
+  /**
+   * Creates matches from the provided data
+   *
+   * @param {FilterProperty<TItem>} property - The property to create matches from
+   * @param {FilterComparisonOperator} comparison - The comparison operator to create matches from
+   * @param {FilterOptionValue[]} values - The values to create matches from
+   * @returns {FilterOptionMatch<TItem>[]} The created matches
+   */
   const matches = useCallback<ListConfigMatchesFn<TItem>>(
-    (
-      property: FilterProperty<TItem>,
-      comparison: FilterComparisonOperator,
-      values: FilterOptionValue[]
-    ): FilterOptionMatch<TItem>[] => {
+    (property, comparison, values) => {
       return values.map((value) => match(property, comparison, value));
     },
     [match]
   );
 
+  /**
+   * Negates a match
+   *
+   * @param {FilterOptionMatch<TItem>} match - The match to negate
+   * @returns {FilterOptionMatch<TItem>} The negated match
+   */
   const not = useCallback<ListConfigNotFn<TItem>>(
-    ({
-      value,
-      property,
-      comparison,
-    }: FilterOptionMatch<TItem>): FilterOptionMatch<TItem> => {
+    ({ property, comparison, value }) => {
       return {
-        value,
         property,
         comparison: invertMatchComparison(comparison),
+        value,
       };
     },
     []
   );
 
-  const option = useCallback<ListConfigOptionFn<TItem>>(
-    (...params: ListConfigOptionFnParams<TItem>): FilterOptionData<TItem> => {
-      const optionData = filterOptionData<TItem>(...params);
-      registerOption(optionData);
-      return optionData;
-    },
+  /**
+   * Creates & registers a filter option from the provided data
+   *
+   * @param {string} label - The label of the generated filter option
+   * @param {FilterOptionMatch<TItem>} matchOrMatches - The match or matches to create the filter option from
+   * @param {FilterOptionConfig} [config = {}] - The config of the generated filter option, gets merged with the shared config
+   * @returns {FilterOptionData<TItem>} The created filter option
+   */
+  const option = useCallback<ListConfigOptionFn<TItem>>((...params) => {
+    const optionData = filterOptionData<TItem>(...params);
+    registerOption(optionData);
+    return optionData;
+  }, []);
+
+  /**
+   * Creates a filter option divider from the provided data
+   *
+   * @param {string} label - The label of the generated filter option divider
+   * @returns {FilterOptionDividerData} The created filter option divider
+   */
+  const divider = useCallback<ListConfigOptionDividerFn>(
+    filterOptionDividerData,
     []
   );
 
-  const divider = useCallback<ListConfigOptionDividerFn>((label: string) => {
-    return filterOptionDividerData(label);
-  }, []);
-
+  /**
+   * Creates & registers filter options from the provided data
+   *
+   * @param {FilterProperty<TItem>} property - The property to create filter options from
+   * @param {FilterComparisonOperator} comparison - The comparison operator to create filter options from
+   * @param {ListConfigOptionLabeledValue[]} labeledValues - The labeled values to create filter options from
+   * @param {FilterOptionConfig} [sharedConfig = {}] - The shared config of the generated filter options, gets merged with the config of each option
+   * @returns {FilterOptionData<TItem>[]} The created filter options
+   */
   const options = useCallback<ListConfigOptionsFn<TItem>>(
-    (
-      property: FilterProperty<TItem>,
-      comparison: FilterComparisonOperator,
-      labeledValues: ListConfigOptionLabeledValue[],
-      sharedConfig: FilterOptionConfig = {}
-    ): FilterOptionData<TItem>[] => {
+    (property, comparison, labeledValues, sharedConfig = {}) => {
       return labeledValues.map(({ label, value, config }) => {
         return option(
           label,
@@ -362,6 +448,14 @@ export const useListConfig: UseListConfig = <TItem extends object>(
     [match, option]
   );
 
+  /**
+   * Creates & registers a filter from the provided data
+   *
+   * @param {string} label - The label of the generated filter
+   * @param {MaybePromise<MaybePromise<FilterOptionDataOrSignature<TItem> | FilterOptionDividerData>[]>} optionsOrSignatures - The options or signatures to create the filter from, can be a promise of an array of promises
+   * @param {FilterConfig} [config = {}] - The config of the generated filter, gets merged with the shared config
+   * @returns {FilterData} The created filter
+   */
   const filter = useCallback<ListConfigFilterFn<TItem>>(
     (
       ...[label, optionsOrSignatures, config]: ListConfigFilterFnParams<TItem>
@@ -374,6 +468,14 @@ export const useListConfig: UseListConfig = <TItem extends object>(
     [hydrateFilterConfig, registerFilter]
   );
 
+  /**
+   * Creates & registers a filter preset from the provided data
+   *
+   * @param {string} label - The label of the generated filter preset
+   * @param {MaybePromise<MaybePromise<FilterOptionDataOrSignature<TItem>>[]>} optionsOrSignatures - The options or signatures to create the filter preset from, can be a promise of an array of promises
+   * @param {FilterPresetConfig} [config = {}] - The config of the generated filter preset, gets merged with the shared config
+   * @returns {FilterPresetData} The created filter preset
+   */
   const filterPreset = useCallback<ListConfigFilterPresetFn<TItem>>(
     (...params: ListConfigFilterPresetFnParams<TItem>): FilterPresetData => {
       const presetData = filterPresetData<TItem>(...params);
@@ -400,6 +502,9 @@ export const useListConfig: UseListConfig = <TItem extends object>(
       writeParams: null,
     });
 
+  /**
+   * A reactive, valid list config object
+   */
   const config = useMemo<ListContextConfig<TItem>>(() => {
     const optionsList = Array.from(optionsMap.values());
     const filtersList = Array.from(filtersMap.values()).sort(