+++ src/components/List/examples/PokemonList.example.tsx
@@ -60,7 +60,7 @@ type PokemonListResponse = {
 };
 
 const DEFAULT_POKEAPI_URL = "https://pokeapi.co/api/v2/pokemon";
-const LIMIT = 100;
+const LIMIT = 250;
 
 const fetchPokemonData = async (): Promise<Pokemon[]> => {
   const fetchDataSubset = async (baseUrl = DEFAULT_POKEAPI_URL) => {
@@ -104,23 +104,20 @@ const usePokemonListConfig = () => {
     configureSearchParams,
   } = useListConfig(usePokemonData);
 
-  // build & register the name filter once
-  const filters = useStatic(() => {
-    configureSearchParams({ sync: "read" });
-    // build & register a filter option for clefairy
+  // make the list's options read the search params
+  useStatic(() => configureSearchParams({ sync: "read" }));
+
+  // declare name options once
+  const names = useStatic(() => {
     const clefairy = option("Clefairy", match("name", "=", "clefairy"), {
       color: "primary",
     });
-
-    // build & register a filter option for pikachu or raichu
     const pikachuOrRaichu = option(
       "Pikachu or Raichu",
       matches("name", "=", ["pikachu", "raichu"]),
-      { color: "pending-dark", operator: BooleanOperators.OR }
+      { color: "pending-dark", operator: "OR" }
     );
-
-    // build & register multiple filter options for entei, cresselia & suicune
-    const others = options(
+    const legendaries = options(
       "name",
       "=",
       [
@@ -128,228 +125,136 @@ const usePokemonListConfig = () => {
         { label: "Cresselia", value: "cresselia" },
         { label: "Suicune", value: "suicune" },
       ],
-      {
-        icon: "Stars",
-        color: "primary",
-      }
-    );
-    // build & register the name filter
-    filter("Name", [clefairy, pikachuOrRaichu, ...others], {
-      multi: true,
-    });
-
-    // declare the base experience filter options asynchronously
-    async.filter(
-      "Base Experience",
-      async.options("base_experience", "<=", async () => {
-        await delay(250);
-        return arrayOf(10, (index): ListConfigOptionLabeledValue => {
-          const exp = index * 50;
-          const label = `${exp} XP`;
-          return { label, value: exp, config: {} };
-        });
-      })
+      { icon: "Stars", color: "primary" }
     );
 
-    // build & register the type filter options
-    const types: ListConfigOptionLabeledValue[] = [
-      {
-        label: "Dark",
-        value: "dark",
-        config: {
-          color: "gray-800",
-          icon: "Moon",
-        },
-      },
-      {
-        label: "Electric",
-        value: "electric",
-        config: {
-          color: "pending",
-          icon: "Lightning",
-          disabled: true,
-        },
-      },
+    const all = [clefairy, pikachuOrRaichu, ...legendaries];
+
+    return {
+      clefairy,
+      pikachuOrRaichu,
+      legendaries,
+      all,
+    };
+  });
+
+  // declare base experience options once
+  const baseExperiences = useStatic(() =>
+    async.options("base_experience", "<=", async () => {
+      await delay(250);
+      return arrayOf(10, (index): ListConfigOptionLabeledValue => {
+        const exp = index * 50;
+        const label = `${exp} XP`;
+        return { label, value: exp, config: {} };
+      });
+    })
+  );
+
+  // declare type options once
+  // pokemon can have up to 2 types, so we need to declare options that match both types
+  const types = useStatic(() => {
+    const typesLabeledValues: ListConfigOptionLabeledValue[] = [
       {
         label: "Fire",
         value: "fire",
-        config: {
-          color: "error",
-          icon: "Fire",
-        },
+        config: { color: "error", icon: "Fire" },
       },
       {
         label: "Grass",
         value: "grass",
-        config: {
-          color: "success",
-          icon: "Flower1",
-        },
+        config: { color: "success", icon: "Flower1" },
       },
       {
         label: "Water",
         value: "water",
-        config: {
-          color: "primary",
-          icon: "Water",
-        },
+        config: { color: "primary", icon: "Water" },
       },
       {
-        label: "Bug",
-        value: "bug",
-        config: {
-          color: "warning-dark",
-          icon: "Bug",
-        },
-      },
-      {
-        label: "Dragon",
-        value: "dragon",
-        config: {
-          color: "warning-dark",
-        },
-      },
-      {
-        label: "Fairy",
-        value: "fairy",
-        config: {
-          color: "primary-medium",
-          icon: "Stars",
-        },
-      },
-      {
-        label: "Ghost",
-        value: "ghost",
-        config: {
-          color: "gray-900",
-          icon: "Snapchat",
-        },
-      },
-      {
-        label: "Steel",
-        value: "steel",
-        config: {
-          color: "gray-400",
-          icon: "Shield",
-        },
-      },
-      {
-        label: "Fighting",
-        value: "fighting",
-        config: {
-          color: "warning",
-          icon: "Diamonds",
-        },
-      },
-      {
-        label: "Normal",
-        value: "normal",
-        config: {
-          color: "gray",
-          icon: "Circle",
-        },
-      },
-      {
-        label: "Psychic",
-        value: "psychic",
-        config: {
-          color: "primary-dark",
-          icon: "PhoneVibrate",
-        },
-      },
-      {
-        label: "Ice",
-        value: "ice",
-        config: {
-          color: "primary",
-          icon: "Snow",
-        },
-      },
-      {
-        label: "Poison",
-        value: "poison",
-        config: {
-          color: "primary-dark",
-          icon: "Cloud",
-          default: true,
-        },
-      },
-      {
-        label: "Flying",
-        value: "flying",
-        config: {
-          color: "gray-400",
-          icon: "Feather",
-        },
+        label: "Electric",
+        value: "electric",
+        config: { color: "pending", icon: "Lightning" },
       },
     ];
+    const typeProperties: FilterProperty<Pokemon>[] = [
+      `types.0.type.name`,
+      `types.1.type.name`,
+    ];
 
-    const typeOptions = types.map((type) => {
-      const properties: FilterProperty<Pokemon>[] = [
-        `types.0.type.name`,
-        `types.1.type.name`,
-      ];
+    const all = typesLabeledValues.map(({ label, value, config }) =>
+      option(
+        label,
+        typeProperties.map((property) => match(property, "=", value)),
+        { ...(config ?? {}), operator: "OR" }
+      )
+    );
+    const [fire, grass, water, electric] = all;
+    return {
+      all,
+      fire,
+      grass,
+      water,
+      electric,
+    };
+  });
 
-      return option(
-        type.label,
-        properties.map((property) => match(property, "=", type.value)),
-        { ...type.config, operator: "OR" }
-      );
-    });
-    // build & register the type filter
-    filter("Type", typeOptions, {
-      operator: BooleanOperators.AND,
-    });
-
-    // build & register a filter option for heavy weight
-
-    const heavyOption = option("Heavy", match("weight", ">=", 200), {
+  // declare weight options once
+  const weights = useStatic(() => {
+    const heavy = option("Heavy", match("weight", ">=", 200), {
       color: "gray-600",
       icon: "BoxArrowDown",
     });
+    const light = option("Light", match("weight", "<", 100), {
+      color: "gray-200",
+      icon: "Feather",
+    });
+    const medium = option(
+      "Medium",
+      [match("weight", ">=", 100), match("weight", "<", 200)],
+      {
+        operator: "AND",
+        color: "gray-400",
+      }
+    );
+    const all = [light, medium, heavy];
+    return {
+      heavy,
+      light,
+      medium,
+      all,
+    };
+  });
 
-    filter("Weight", [
-      // build & register a filter option for light weight
-      option("Light", match("weight", "<", 100), {
-        color: "gray-200",
-        icon: "Feather",
-      }),
-      // build & register a filter option for medium weight
-      option(
-        "Medium",
-        [match("weight", ">=", 100), match("weight", "<", 200)],
-        { operator: BooleanOperators.AND, color: "gray-400" }
-      ),
-      heavyOption,
-    ]);
+  // declare filters once
+  // order matters.
+  useStatic(() => {
+    filter("Name", names.all, { multi: true });
+    async.filter("Base Experience", baseExperiences);
+    filter("Type", types.all, { operator: BooleanOperators.OR });
+    filter("Weight", weights.all);
+  });
 
-    // declare some filter presets
+  // declare some filter presets once
+  // order matters.
+  useStatic(() => {
     filterPreset(
-      "Fire Birds",
-      typeOptions.filter(
-        (option) => option.label === "Fire" || option.label === "Flying"
-      ),
+      "PyroElectric lightweights",
+      [types.fire, types.electric, weights.light],
       {
         color: "error",
       }
     );
-
-    const electricType = typeOptions.find(
-      (option) => option.label === "Electric"
-    );
-
     filterPreset(
       "Electric Pikachu or Raichu",
-      electricType ? [electricType, pikachuOrRaichu] : [pikachuOrRaichu],
+      [types.electric, names.pikachuOrRaichu],
       {
         color: "pending",
         exclusive: true,
       }
     );
-
-    filterPreset("Obese mice", [pikachuOrRaichu, heavyOption]);
+    filterPreset("Obese mice", [names.pikachuOrRaichu, weights.heavy]);
   });
-  return { config, filters };
+
+  return config;
 };
 
 const LoadingFill = styled(Loading)`
@@ -405,7 +310,7 @@ const Renderer = () => {
 };
 
 export const PokemonListExample = () => {
-  const { config } = usePokemonListConfig();
+  const config = usePokemonListConfig();
 
   return (
     <ListContextProvider config={config}>