+++ src/components/Text/Text.component.tsx
@@ -2,7 +2,12 @@ import { styled } from "styled-components";
 
 import { buildTypographyStyle, defaultTypographyProps } from "../../typography";
 
-import { useClassName, useStyleProps, useTestId } from "@utils";
+import {
+  useClassName,
+  useHtmlAttribute,
+  useStyleProps,
+  useTestId,
+} from "@utils";
 
 import type { StyleProps, TestIdProps, TextProps } from "@types";
 
@@ -23,8 +28,14 @@ const Text = (props: TextProps & TestIdProps): JSX.Element => {
   const innerProps = useStyleProps(props);
   const testId = useTestId("text", props);
   const className = useClassName(props);
+  const id = useHtmlAttribute(props.id);
   return (
-    <TextInner data-testid={testId} className={className} {...innerProps}>
+    <TextInner
+      data-testid={testId}
+      className={className}
+      {...innerProps}
+      id={id}
+    >
       {props.children}
     </TextInner>
   );