+++ src/types/typography/typography.types.ts
@@ -2,6 +2,7 @@ import { texts } from "@ubloimmo/front-tokens";
 
 import type { StyleOverrideProps } from "@/types/global/styles.types";
 import type { PaletteColor } from "@/types/themes/palette";
+import type { Nullable } from "@ubloimmo/front-util";
 import type { ReactNode } from "react";
 
 export type TypographyTokens = typeof texts;
@@ -126,6 +127,12 @@ export type TypographyProps = {
    * @default undefined
    */
   children?: ReactNode;
+  /**
+   * A unique identifier for the text / heading component.
+   *
+   * @default undefined
+   */
+  id?: Nullable<string>;
 } & StyleOverrideProps;
 
 export type TextProps = TypographyProps & {